refactor(rust): adjust the kafka services api
